+++ bobashare/src/serde/tests/deserialize/v0.rs
@@ -6,7 +6,7 @@ use crate::serde::{tests::constants, UploadMetadata};
 fn deserialize_upload_with_expiry() {
     let output = UploadMetadata::into_migrated_upload(
         constants::upload_id(),
-        serde_json::from_str(constants::EXAMPLE_UPLOADV1_WITH_EXPIRY_SERIALIZED).unwrap(),
+        serde_json::from_str(constants::EXAMPLE_UPLOADV0_WITH_EXPIRY_SERIALIZED).unwrap(),
     )
     .unwrap();
 
@@ -17,7 +17,7 @@ fn deserialize_upload_with_expiry() {
 fn deserialize_upload_no_expiry() {
     let output = UploadMetadata::into_migrated_upload(
         constants::upload_id(),
-        serde_json::from_str(constants::EXAMPLE_UPLOADV1_NO_EXPIRY_SERIALIZED).unwrap(),
+        serde_json::from_str(constants::EXAMPLE_UPLOADV0_NO_EXPIRY_SERIALIZED).unwrap(),
     )
     .unwrap();